--- PULL_REQUEST_TEMPLATE.md
@@ -1,21 +1,0 @@
-# [{card_number}] {title}
-
-제목 예시: [17] 메인페이지 UI
-
-## 노션 칸반 링크
-
-- [카드 제목](url주소)
-
-## 카드에서 구현 혹은 해결하려는 내용
-
--
--
-
-## 테스트 방법
-
--
--
-
-## 기타 사항
-
--
