+++ src/index.html
@@ -52,7 +52,7 @@
 
 					<div class="game-explain">
 						<img class="audio-image" src="./assets/images/audio/loud.png" alt="soft voice image" />
-						Speck
+						Speak
 						<span class="red-text">LOUDER</span>, jump <span class="red-text">Higher</span>
 						<img
 							class="arrow-jump"
